requirements.txt
streamlit>=1.31.0,<2.0.0
pydantic-ai>=0.0.8
pydantic>=2.5.0,<3.0.0
httpx>=0.25.0,<1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0,<4.0.0
sentence-transformers>=2.2.0,<3.0.0
faiss-cpu>=1.7.4,<2.0.0
openai>=1.12.0,<2.0.0
numpy>=1.24.0,<2.0.0
tiktoken>=0.5.0,<1.0.0
